+++ backend/src/main/java/io/f1/backend/domain/game/app/RoomService.java
@@ -1,7 +1,12 @@
 package io.f1.backend.domain.game.app;
 
-import io.f1.backend.domain.game.store.RoomRepository;
+import static io.f1.backend.domain.game.mapper.RoomMapper.toRoomSetting;
+
 import io.f1.backend.domain.game.dto.request.RoomCreateRequest;
+import io.f1.backend.domain.game.model.GameSetting;
+import io.f1.backend.domain.game.model.Player;
+import io.f1.backend.domain.game.model.RoomSetting;
+import io.f1.backend.domain.game.store.RoomRepository;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,8 +17,15 @@ public class RoomService {
 
     private final RoomRepository roomRepository;
 
-    public Long saveRoom(RoomCreateRequest request,Map<String, Object> loginUser) {
-        return roomRepository.saveRoom(request,loginUser);
+    public Long saveRoom(RoomCreateRequest request, Map<String, Object> loginUser) {
+
+        //todo 제일 작은 index quizId 가져와서 gameSetting
+        GameSetting gameSetting = new GameSetting(1L, 10, 60);
+        //todo security에서 가져오는걸로 변경
+        Player host = new Player((Long) loginUser.get("id"), loginUser.get("nickname").toString());
+        RoomSetting roomSetting = toRoomSetting(request);
+
+        return roomRepository.saveRoom(gameSetting, host, roomSetting);
     }
 
 }
